\name{simpact.getconfig}
\alias{simpact.getconfig}
\title{simpact.getconfig}
\description{
	TODO
}
\usage{
 	simpact.getconfig(configParams, show = FALSE)
}
\arguments{
	\item{configParams}{}
	\item{show}{}
}
\details{
	TODO
}
\value{
 	TODO
}

\examples{
}

\keyword{manip}

