\name{simpact.run.direct}
\alias{simpact.run.direct}
\title{simpact.run.direct}
\description{
	TODO
}
\usage{
	simpact.run.direct(configFile, outputFile = NULL, release = TRUE, slowalg = FALSE, parallel = FALSE, seed = -1, destDir=NULL)
}
\arguments{
	\item{configFile}{}
	\item{outputFile}{}
	\item{release}{}
	\item{slowalg}{}
	\item{parallel}{}
	\item{seed}{}
	\item{destDir}{}
}
\details{
	TODO
}
\value{
 	None.
}

\examples{
	TODO
}

\keyword{ simpact }

