\name{simpact.run}
\alias{simpact.run}
\title{simpact.run}
\description{
	TODO
}
\usage{
 	simpact.run(configParams, destDir, agedist = simpact.sa2003, intervention = NULL, release = TRUE, slowalg = FALSE, parallel=FALSE, seed=-1, dryrun = FALSE)
}
\arguments{
	\item{configParams}{}
	\item{destDir}{}
	\item{agedist}{}
	\item{intervention}{}
	\item{release}{}
	\item{slowalg}{}
	\item{parallel}{}
	\item{seed}{}
	\item{dryrun}{}
}
\details{
	TODO
}
\value{
 	TODO
}

\examples{
	TODO
}

\keyword{ simpact }

